+++ physicalspeaker.h
@@ -9,8 +9,8 @@ class PhysicalSpeaker {
 
   virtual void begin() = 0;
 
-  virtual void toggle(uint32_t c) = 0;
-  virtual void maintainSpeaker(uint32_t c, uint64_t microseconds) = 0;
+  virtual void toggle(int64_t c) = 0;
+  virtual void maintainSpeaker(int64_t c, uint64_t microseconds) = 0;
   virtual void beginMixing() = 0;
   virtual void mixOutput(uint8_t v) = 0;
